refactor(events): component style

+++ events/frontend/src/components/event/EventItem/EventItem.jsx
@@ -5,13 +5,10 @@ import classes from './EventItem.module.css';
 export function EventItem({ event }) {
   const submit = useSubmit();
 
-  function startDeleteHandler() {
+  const startDeleteHandler = () => {
     const proceed = window.confirm('Are you sure?');
-
-    if (proceed) {
-      submit(null, { method: 'delete' });
-    }
-  }
+    if (proceed) submit(null, { method: 'delete' });
+  };
 
   return (
     <article className={classes.event}>
